+++ packages/webview_flutter/webview_flutter_android/test/android_webview_test.mocks.dart
@@ -2,15 +2,15 @@
 // in webview_flutter_android/test/android_webview_test.dart.
 // Do not manually edit this file.
 
-import 'dart:async' as _i4;
-import 'dart:typed_data' as _i6;
-import 'dart:ui' as _i7;
+import 'dart:async' as _i5;
+import 'dart:typed_data' as _i7;
+import 'dart:ui' as _i4;
 
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:webview_flutter_android/src/android_webview.dart' as _i2;
 import 'package:webview_flutter_android/src/android_webview.pigeon.dart' as _i3;
 
-import 'test_android_webview.pigeon.dart' as _i5;
+import 'test_android_webview.pigeon.dart' as _i6;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -28,13 +28,17 @@ class _FakeDownloadListener_0 extends _i1.Fake
 class _FakeJavaScriptChannel_1 extends _i1.Fake
     implements _i2.JavaScriptChannel {}
 
-class _FakeWebChromeClient_2 extends _i1.Fake implements _i2.WebChromeClient {}
+class _FakeWebViewPoint_2 extends _i1.Fake implements _i3.WebViewPoint {}
 
-class _FakeWebSettings_3 extends _i1.Fake implements _i2.WebSettings {}
+class _FakeWebChromeClient_3 extends _i1.Fake implements _i2.WebChromeClient {}
 
-class _FakeWebView_4 extends _i1.Fake implements _i2.WebView {}
+class _FakeWebSettings_4 extends _i1.Fake implements _i2.WebSettings {}
 
-class _FakeWebViewClient_5 extends _i1.Fake implements _i2.WebViewClient {}
+class _FakeOffset_5 extends _i1.Fake implements _i4.Offset {}
+
+class _FakeWebView_6 extends _i1.Fake implements _i2.WebView {}
+
+class _FakeWebViewClient_7 extends _i1.Fake implements _i2.WebViewClient {}
 
 /// A class which mocks [CookieManagerHostApi].
 ///
@@ -46,14 +50,14 @@ class MockCookieManagerHostApi extends _i1.Mock
   }
 
   @override
-  _i4.Future<bool> clearCookies() =>
+  _i5.Future<bool> clearCookies() =>
       (super.noSuchMethod(Invocation.method(#clearCookies, []),
-          returnValue: Future<bool>.value(false)) as _i4.Future<bool>);
+          returnValue: Future<bool>.value(false)) as _i5.Future<bool>);
   @override
-  _i4.Future<void> setCookie(String? arg_url, String? arg_value) =>
+  _i5.Future<void> setCookie(String? arg_url, String? arg_value) =>
       (super.noSuchMethod(Invocation.method(#setCookie, [arg_url, arg_value]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
 }
 
 /// A class which mocks [DownloadListener].
@@ -103,7 +107,7 @@ class MockJavaScriptChannel extends _i1.Mock implements _i2.JavaScriptChannel {
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestDownloadListenerHostApi extends _i1.Mock
-    implements _i5.TestDownloadListenerHostApi {
+    implements _i6.TestDownloadListenerHostApi {
   MockTestDownloadListenerHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -118,7 +122,7 @@ class MockTestDownloadListenerHostApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestJavaObjectHostApi extends _i1.Mock
-    implements _i5.TestJavaObjectHostApi {
+    implements _i6.TestJavaObjectHostApi {
   MockTestJavaObjectHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -133,7 +137,7 @@ class MockTestJavaObjectHostApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestJavaScriptChannelHostApi extends _i1.Mock
-    implements _i5.TestJavaScriptChannelHostApi {
+    implements _i6.TestJavaScriptChannelHostApi {
   MockTestJavaScriptChannelHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -148,7 +152,7 @@ class MockTestJavaScriptChannelHostApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestWebChromeClientHostApi extends _i1.Mock
-    implements _i5.TestWebChromeClientHostApi {
+    implements _i6.TestWebChromeClientHostApi {
   MockTestWebChromeClientHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -164,7 +168,7 @@ class MockTestWebChromeClientHostApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestWebSettingsHostApi extends _i1.Mock
-    implements _i5.TestWebSettingsHostApi {
+    implements _i6.TestWebSettingsHostApi {
   MockTestWebSettingsHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -240,7 +244,7 @@ class MockTestWebSettingsHostApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestWebStorageHostApi extends _i1.Mock
-    implements _i5.TestWebStorageHostApi {
+    implements _i6.TestWebStorageHostApi {
   MockTestWebStorageHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -259,7 +263,7 @@ class MockTestWebStorageHostApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestWebViewClientHostApi extends _i1.Mock
-    implements _i5.TestWebViewClientHostApi {
+    implements _i6.TestWebViewClientHostApi {
   MockTestWebViewClientHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -275,7 +279,7 @@ class MockTestWebViewClientHostApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestWebViewHostApi extends _i1.Mock
-    implements _i5.TestWebViewHostApi {
+    implements _i6.TestWebViewHostApi {
   MockTestWebViewHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -308,7 +312,7 @@ class MockTestWebViewHostApi extends _i1.Mock
           Invocation.method(#loadUrl, [instanceId, url, headers]),
           returnValueForMissingStub: null);
   @override
-  void postUrl(int? instanceId, String? url, _i6.Uint8List? data) =>
+  void postUrl(int? instanceId, String? url, _i7.Uint8List? data) =>
       super.noSuchMethod(Invocation.method(#postUrl, [instanceId, url, data]),
           returnValueForMissingStub: null);
   @override
@@ -340,12 +344,12 @@ class MockTestWebViewHostApi extends _i1.Mock
           Invocation.method(#clearCache, [instanceId, includeDiskFiles]),
           returnValueForMissingStub: null);
   @override
-  _i4.Future<String?> evaluateJavascript(
+  _i5.Future<String?> evaluateJavascript(
           int? instanceId, String? javascriptString) =>
       (super.noSuchMethod(
           Invocation.method(
               #evaluateJavascript, [instanceId, javascriptString]),
-          returnValue: Future<String?>.value()) as _i4.Future<String?>);
+          returnValue: Future<String?>.value()) as _i5.Future<String?>);
   @override
   String? getTitle(int? instanceId) =>
       (super.noSuchMethod(Invocation.method(#getTitle, [instanceId]))
@@ -366,6 +370,10 @@ class MockTestWebViewHostApi extends _i1.Mock
   int getScrollY(int? instanceId) =>
       (super.noSuchMethod(Invocation.method(#getScrollY, [instanceId]),
           returnValue: 0) as int);
+  @override
+  _i3.WebViewPoint getScrollPosition(int? instanceId) =>
+      (super.noSuchMethod(Invocation.method(#getScrollPosition, [instanceId]),
+          returnValue: _FakeWebViewPoint_2()) as _i3.WebViewPoint);
   @override
   void setWebContentsDebuggingEnabled(bool? enabled) => super.noSuchMethod(
       Invocation.method(#setWebContentsDebuggingEnabled, [enabled]),
@@ -412,7 +420,7 @@ class MockTestWebViewHostApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestAssetManagerHostApi extends _i1.Mock
-    implements _i5.TestAssetManagerHostApi {
+    implements _i6.TestAssetManagerHostApi {
   MockTestAssetManagerHostApi() {
     _i1.throwOnMissingStub(this);
   }
@@ -442,7 +450,7 @@ class MockWebChromeClient extends _i1.Mock implements _i2.WebChromeClient {
   @override
   _i2.WebChromeClient copy() =>
       (super.noSuchMethod(Invocation.method(#copy, []),
-          returnValue: _FakeWebChromeClient_2()) as _i2.WebChromeClient);
+          returnValue: _FakeWebChromeClient_3()) as _i2.WebChromeClient);
 }
 
 /// A class which mocks [WebView].
@@ -460,17 +468,17 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
   @override
   _i2.WebSettings get settings =>
       (super.noSuchMethod(Invocation.getter(#settings),
-          returnValue: _FakeWebSettings_3()) as _i2.WebSettings);
+          returnValue: _FakeWebSettings_4()) as _i2.WebSettings);
   @override
-  _i4.Future<void> loadData(
+  _i5.Future<void> loadData(
           {String? data, String? mimeType, String? encoding}) =>
       (super.noSuchMethod(
           Invocation.method(#loadData, [],
               {#data: data, #mimeType: mimeType, #encoding: encoding}),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> loadDataWithBaseUrl(
+  _i5.Future<void> loadDataWithBaseUrl(
           {String? baseUrl,
           String? data,
           String? mimeType,
@@ -485,117 +493,122 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
             #historyUrl: historyUrl
           }),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> loadUrl(String? url, Map<String, String>? headers) =>
+  _i5.Future<void> loadUrl(String? url, Map<String, String>? headers) =>
       (super.noSuchMethod(Invocation.method(#loadUrl, [url, headers]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> postUrl(String? url, _i6.Uint8List? data) =>
+  _i5.Future<void> postUrl(String? url, _i7.Uint8List? data) =>
       (super.noSuchMethod(Invocation.method(#postUrl, [url, data]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<String?> getUrl() =>
+  _i5.Future<String?> getUrl() =>
       (super.noSuchMethod(Invocation.method(#getUrl, []),
-          returnValue: Future<String?>.value()) as _i4.Future<String?>);
+          returnValue: Future<String?>.value()) as _i5.Future<String?>);
   @override
-  _i4.Future<bool> canGoBack() =>
+  _i5.Future<bool> canGoBack() =>
       (super.noSuchMethod(Invocation.method(#canGoBack, []),
-          returnValue: Future<bool>.value(false)) as _i4.Future<bool>);
+          returnValue: Future<bool>.value(false)) as _i5.Future<bool>);
   @override
-  _i4.Future<bool> canGoForward() =>
+  _i5.Future<bool> canGoForward() =>
       (super.noSuchMethod(Invocation.method(#canGoForward, []),
-          returnValue: Future<bool>.value(false)) as _i4.Future<bool>);
+          returnValue: Future<bool>.value(false)) as _i5.Future<bool>);
   @override
-  _i4.Future<void> goBack() =>
+  _i5.Future<void> goBack() =>
       (super.noSuchMethod(Invocation.method(#goBack, []),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> goForward() =>
+  _i5.Future<void> goForward() =>
       (super.noSuchMethod(Invocation.method(#goForward, []),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> reload() =>
+  _i5.Future<void> reload() =>
       (super.noSuchMethod(Invocation.method(#reload, []),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> clearCache(bool? includeDiskFiles) =>
+  _i5.Future<void> clearCache(bool? includeDiskFiles) =>
       (super.noSuchMethod(Invocation.method(#clearCache, [includeDiskFiles]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<String?> evaluateJavascript(String? javascriptString) => (super
+  _i5.Future<String?> evaluateJavascript(String? javascriptString) => (super
       .noSuchMethod(Invocation.method(#evaluateJavascript, [javascriptString]),
-          returnValue: Future<String?>.value()) as _i4.Future<String?>);
+          returnValue: Future<String?>.value()) as _i5.Future<String?>);
   @override
-  _i4.Future<String?> getTitle() =>
+  _i5.Future<String?> getTitle() =>
       (super.noSuchMethod(Invocation.method(#getTitle, []),
-          returnValue: Future<String?>.value()) as _i4.Future<String?>);
+          returnValue: Future<String?>.value()) as _i5.Future<String?>);
   @override
-  _i4.Future<void> scrollTo(int? x, int? y) =>
+  _i5.Future<void> scrollTo(int? x, int? y) =>
       (super.noSuchMethod(Invocation.method(#scrollTo, [x, y]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> scrollBy(int? x, int? y) =>
+  _i5.Future<void> scrollBy(int? x, int? y) =>
       (super.noSuchMethod(Invocation.method(#scrollBy, [x, y]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<int> getScrollX() =>
+  _i5.Future<int> getScrollX() =>
       (super.noSuchMethod(Invocation.method(#getScrollX, []),
-          returnValue: Future<int>.value(0)) as _i4.Future<int>);
+          returnValue: Future<int>.value(0)) as _i5.Future<int>);
   @override
-  _i4.Future<int> getScrollY() =>
+  _i5.Future<int> getScrollY() =>
       (super.noSuchMethod(Invocation.method(#getScrollY, []),
-          returnValue: Future<int>.value(0)) as _i4.Future<int>);
+          returnValue: Future<int>.value(0)) as _i5.Future<int>);
   @override
-  _i4.Future<void> setWebViewClient(_i2.WebViewClient? webViewClient) =>
+  _i5.Future<_i4.Offset> getScrollPosition() =>
+      (super.noSuchMethod(Invocation.method(#getScrollPosition, []),
+              returnValue: Future<_i4.Offset>.value(_FakeOffset_5()))
+          as _i5.Future<_i4.Offset>);
+  @override
+  _i5.Future<void> setWebViewClient(_i2.WebViewClient? webViewClient) =>
       (super.noSuchMethod(Invocation.method(#setWebViewClient, [webViewClient]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> addJavaScriptChannel(
+  _i5.Future<void> addJavaScriptChannel(
           _i2.JavaScriptChannel? javaScriptChannel) =>
       (super.noSuchMethod(
           Invocation.method(#addJavaScriptChannel, [javaScriptChannel]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> removeJavaScriptChannel(
+  _i5.Future<void> removeJavaScriptChannel(
           _i2.JavaScriptChannel? javaScriptChannel) =>
       (super.noSuchMethod(
           Invocation.method(#removeJavaScriptChannel, [javaScriptChannel]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> setDownloadListener(_i2.DownloadListener? listener) =>
+  _i5.Future<void> setDownloadListener(_i2.DownloadListener? listener) =>
       (super.noSuchMethod(Invocation.method(#setDownloadListener, [listener]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> setWebChromeClient(_i2.WebChromeClient? client) =>
+  _i5.Future<void> setWebChromeClient(_i2.WebChromeClient? client) =>
       (super.noSuchMethod(Invocation.method(#setWebChromeClient, [client]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> setBackgroundColor(_i7.Color? color) =>
+  _i5.Future<void> setBackgroundColor(_i4.Color? color) =>
       (super.noSuchMethod(Invocation.method(#setBackgroundColor, [color]),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
-  _i4.Future<void> release() =>
+  _i5.Future<void> release() =>
       (super.noSuchMethod(Invocation.method(#release, []),
           returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i4.Future<void>);
+          returnValueForMissingStub: Future<void>.value()) as _i5.Future<void>);
   @override
   _i2.WebView copy() => (super.noSuchMethod(Invocation.method(#copy, []),
-      returnValue: _FakeWebView_4()) as _i2.WebView);
+      returnValue: _FakeWebView_6()) as _i2.WebView);
 }
 
 /// A class which mocks [WebViewClient].
@@ -641,5 +654,5 @@ class MockWebViewClient extends _i1.Mock implements _i2.WebViewClient {
           returnValueForMissingStub: null);
   @override
   _i2.WebViewClient copy() => (super.noSuchMethod(Invocation.method(#copy, []),
-      returnValue: _FakeWebViewClient_5()) as _i2.WebViewClient);
+      returnValue: _FakeWebViewClient_7()) as _i2.WebViewClient);
 }